requirements-api.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP clients
httpx==0.26.0
requests==2.31.0

# GCP clients
google-cloud-pubsub==2.19.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
